Fixed Material Bug & Added Tutorial
This commit accomplishes two tasks which include fixing the material bug as well as adding a tutorial on how to move the graph. To fix the bug I added a wait so the asynchronous task of predicting the flower can finish before going into setting the new point. The tutorial includes two text boxes that describe how to move the graph.